Expose makeResponse
