+++ Sources/JWTKeychain/Support/TokenGenerators.swift
@@ -1,4 +1,5 @@
 import JWTProvider
+import Vapor
 
 public struct TokenGenerators {
     internal let apiAccess: ExpireableSigner
@@ -41,4 +42,60 @@ extension TokenGenerators {
             signerMap: signerMap
         )
     }
+}
+
+
+// MARK: Helper
+
+public extension TokenGenerators {
+    public func makeResponse<U: JWTKeychainUser>(
+        for user: U,
+        withOptions responseOptions: ResponseOptions
+    ) throws -> ResponseRepresentable {
+        var response = JSON()
+
+        if responseOptions.contains(.access) {
+            try response.set(
+                "accessToken",
+                self
+                    .apiAccessTokenGenerator
+                    .generateToken(for: user)
+                    .string
+            )
+        }
+        if
+            responseOptions.contains(.refresh),
+            let refreshTokenGenerator = self.refreshTokenGenerator
+        {
+            try response.set(
+                "refreshToken",
+                refreshTokenGenerator.generateToken(for: user).string
+            )
+        }
+        if responseOptions.contains(.user) {
+            if responseOptions == [.user] {
+                // make an exception when only user is to be returned
+                // -> return user as root level object
+                return try user.makeJSON()
+            } else {
+                try response.set("user", user)
+            }
+        }
+
+        return response
+    }
+}
+
+public struct ResponseOptions: OptionSet {
+    public let rawValue: Int
+
+    public init(rawValue: ResponseOptions.RawValue) {
+        self.rawValue = rawValue
+    }
+
+    public static let access = ResponseOptions(rawValue: 1 << 0)
+    public static let refresh = ResponseOptions(rawValue: 1 << 1)
+    public static let user = ResponseOptions(rawValue: 1 << 2)
+
+    public static let all: ResponseOptions = [.access, .refresh, .user]
 }
